Criação de um dialog de confirmação de exclusão

+++ src/app/courses/containers/courses/courses.component.ts
@@ -1,20 +1,20 @@
 import { CoursesService } from '../../services/courses.service';
 import { Component, OnInit } from '@angular/core';
 import { Course } from '../../model/course';
-import { Observable,  of } from 'rxjs';
+import { Observable, of } from 'rxjs';
 import { catchError } from 'rxjs/operators';
 import { ErrorDialogComponent } from 'src/app/shared/components/error-dialog/error-dialog.component';
 import { ActivatedRoute, Router } from '@angular/router';
 import { MatDialog } from '@angular/material/dialog';
 import { MatSnackBar } from '@angular/material/snack-bar';
+import { ConfirmationDialogComponent } from 'src/app/shared/components/confirmation-dialog/confirmation-dialog.component';
 
 @Component({
   selector: 'app-courses',
   templateUrl: './courses.component.html',
-  styleUrls: ['./courses.component.scss']
+  styleUrls: ['./courses.component.scss'],
 })
 export class CoursesComponent implements OnInit {
-
   courses$: Observable<Course[]> | null = null;
 
   displayedColumns = ['name', 'category', 'actions'];
@@ -25,49 +25,55 @@ export class CoursesComponent implements OnInit {
     private router: Router,
     private route: ActivatedRoute,
     private snackBar: MatSnackBar
-    ) {
+  ) {
     this.refresh();
   }
 
   refresh() {
-    this.courses$ = this.coursesService.list()
-    .pipe(
-      catchError(error => {
+    this.courses$ = this.coursesService.list().pipe(
+      catchError((error) => {
         this.onError('Erro ao carregar cursos.');
-        return of([])
+        return of([]);
       })
     );
   }
 
   onError(errorMsg: string) {
     this.dialog.open(ErrorDialogComponent, {
-      data: errorMsg
+      data: errorMsg,
     });
   }
 
-  ngOnInit(): void {
+  ngOnInit(): void {}
+
+  onAdd() {
+    this.router.navigate(['new'], { relativeTo: this.route });
   }
 
-  onAdd(){
-    this.router.navigate(['new'], {relativeTo: this.route});
-  }
-
-  onEdit(course: Course){
+  onEdit(course: Course) {
     // console.log('CoursesComponent >>> onEdit >> ' + course);
-    this.router.navigate(['edit', course._id], {relativeTo: this.route});
+    this.router.navigate(['edit', course._id], { relativeTo: this.route });
   }
 
-  onRemove(course: Course){
-    this.coursesService.remove(course._id).subscribe(
-      () => {
-        this.snackBar.open('Curso removido com sucesso!', 'X', {
-          duration: 5000,
-          verticalPosition: 'top',
-          horizontalPosition: 'center'
-        });
-        this.refresh();
-      },
-      () => this.onError('Erro ao tentar remover curso.')
-    );
+  onRemove(course: Course) {
+    const dialogRef = this.dialog.open(ConfirmationDialogComponent, {
+      data: 'Tem certeza que deseja remover este curso?',
+    });
+
+    dialogRef.afterClosed().subscribe((result: boolean) => {
+      if (result) {
+        this.coursesService.remove(course._id).subscribe(
+          () => {
+            this.snackBar.open('Curso removido com sucesso!', 'X', {
+              duration: 5000,
+              verticalPosition: 'top',
+              horizontalPosition: 'center',
+            });
+            this.refresh();
+          },
+          () => this.onError('Erro ao tentar remover curso.')
+        );
+      }
+    });
   }
 }
